requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
sqladmin>=0.20.0
httpx>=0.28.0
cryptography>=44.0.0
python-multipart>=0.0.18
itsdangerous>=2.2.0
psycopg2-binary>=2.9.9
PyJWT>=2.9.0
slowapi>=0.1.9
redis>=5.0.0
